Moved template to a separate folder and fixed paths
