+++ client/src/components/Templates/templates.js
@@ -1,11 +1,11 @@
 import React, { Component } from "react";
-import Sidebar from "./subComponents/sidebar";
-import Navbar from "./subComponents/navbar";
-import "./templates/tempTemplate1.png";
-import "./templates/tempTemplate2.png";
-import "./templates/tempTemplate3.png";
+import Sidebar from "../subComponents/sidebar";
+import Navbar from "../subComponents/navbar";
+import "./tempTemplate1.png";
+import "./tempTemplate2.png";
+import "./tempTemplate3.png";
 import { Link } from "react-router-dom";
-import "./CSS/templates.css";
+import "./templates.css";
 
 class Templates extends Component {
   render() {
@@ -31,12 +31,12 @@ class Templates extends Component {
                     color: "black"
                   }}
                   to={{
-                    pathname: "/templates/template-1", // component being Linked to
+                    pathname: "/template-1", // component being Linked to
                     state: { templateIndex: false } // Setting Index passed to template- false means new
                   }}
                 >
                   <img
-                    src={require("./templates/tempTemplate.png")} //import pics
+                    src={require("./tempTemplate.png")} //import pics
                     alt=""
                     height="100rem"
                   />
@@ -49,12 +49,12 @@ class Templates extends Component {
                     color: "black"
                   }}
                   to={{
-                    pathname: "/templates/template-2", // component being Linked to
+                    pathname: "/template-2", // component being Linked to
                     state: { templateIndex: false } // Setting Index passed to template- false means new
                   }}
                 >
                   <img
-                    src={require("./templates/tempTemplate1.png")} //import pics
+                    src={require("./tempTemplate1.png")} //import pics
                     alt=""
                     height="100rem"
                   />
@@ -67,12 +67,12 @@ class Templates extends Component {
                     color: "black"
                   }}
                   to={{
-                    pathname: "/templates/template-3", // component being Linked to
+                    pathname: "/template-3", // component being Linked to
                     state: { templateIndex: false } // Setting Index passed to template- false means new
                   }}
                 >
                   <img
-                    src={require("./templates/tempTemplate2.png")} //import pics
+                    src={require("./tempTemplate2.png")} //import pics
                     alt=""
                     height="100rem"
                   />
